we aren't going to run late INIT blocks, so don't save them
I was looking over #1674, where it was suggested that late INIT blocks
behave like BEGIN, and Larry approved of that (25 years ago).

But history has moved on, we have 25 years of code developed under the
current behaviour of INIT and I suspect such a change is more likely
to break existing code than fix anything.

But looking at the code in S_process_special_blocks() we do still push
the CV onto PL_initav, even though that CV will never be called.

So don't push the CV if it's too late to call it.