Deparse: fix 3-arg susbtr_left deparsing
v5.41.7-43-gcdbed2a40e introduced the OP_SUBSTR_LEFT op,
which is an optimised version of OP_SUBSTR for when the offset
is zero and the replacement string is missing or ''.

Unfortunately the deparsing for this OP missed out the closing
parenthesis when the replacement string wasn't present; i.e.:

    substr($lex, 0, 4);

deparsed as:

    substr($lex, 0, 4;

The fix is trivial.