podcheck.t: Minimize exceptions db entries
This program examines the files in the distribution looking for ones
that contain pod.  When one is found, it knows that it is safe to link
to its NAME.

It also maintains a data base that contains two types of entries: 1)
known issues; and 2) entries that are safe to link to that are external
to the distribution.  The latter type includes things like man pages or
perl modules not shipped with the core.

Commit 1fb05c6f7dcd4efde7ac9e3796231129ae970f7a added code to better
know that it was safe to link to files containing pod in the
distribution.  But what happened was that that commit caused all such to
be needlessly written out to the data base.  That db is supposed to
contain things that are exceptions, that can't be found otherwise.  It
is not intended to have the many entries that get resolved in the normal
course of executing this program.

This commit solves this by marking the entries it finds normally in
parsing as to not be written out to the data base.  The data base is
regenerated, removing these unnecessary entries that got added in
1fb05c6f7dcd4efde7ac9e3796231129ae970f7a.