Implement assigning xor (^^=) operator
When I added '^^' I forgot to implement or test the assigning version of
it.

Also it seems `pp_xor` had the left and right arguments round the wrong
way; but until the asymmetry introduced by this change nobody had
noticed before. This is now fixed.

Also adds `B::Deparse` support for the new assigning xor operator