op.c: Make ifdef structure of argument to S_op_clear_gv() less objectionable to syntax parsers
The prior structure of this syntax made it look to tree-sitter-based
parsers as if two separate functions were being defined, though only the
second of which was ever finished. This resulted in the entire rest of
the file appearing to be nested inside the first function, meaning no
top-level function analysis was performed on them. This has various
knock-on effects on code analysis in tooling such as editors or
browsers.

The fix for it moves the `#ifdef` syntax inside the argument parentheses
to ensure that only one function appears to be created, so the entire
rest of the file is parsed as normal.