perlfunc: Remove self-referential links
It is not good form to have links in the middle of a section go back to
the top of that very section.  Someone clicks on it, thinking they are
going to get more detail or a different explanation, and all that
happens is they lose their place in the section.