sv.h: Make SvIsUV() return 1 (rather than 0x80000000) for true condition
This would make it easier to receive the result with `bool` type.

Note that (implicit) casting to C99 `_Bool` automatically converts
any nonzero scalar values to 1, so that this change is technically
not necessary in C99 which perl now requires to compile.  But I think
it will keep the code less surprising.