perlre: Simplify some text
By adding a word to a =heading, a sentence can be removed, and is
clearer.

But since some pod somewhere  may link to that heading, a section at the
end is added with the old name, and pointing to the new one.

Dan Book searched and found no instances in CPAN of the old heading
being linked to, but this guarantees that nothing breaks.