uni/method.t: Update, and don't test specific garbage
The modern way to change UTF-8 to its component bytes is to use
utf8::encode

Some of the tests are making sure that those component bytes aren't
mistaken for being UTF-8.  What those component bytes are is not
actually relevant, but the tests were looking at the specific expected
values of them.  The problem is that these differ on EBCDIC vs ASCII
platforms.  Several commits had been added to try to get the correct
values on both types, but EBCDIC still was getting failures.  And, there
is no need to test for the specific values of these irrelevant bytes.
What is important is that they were not misinterpreted as if they were
UTF-8.

This commit goes back to the original tests before those other commits
were added, and changes the matching pattern to not look for the
specific irrelevant byte values.

Doing so makes the tests pass on both types of platforms.