Allow z/OS to work in either EBCDIC or ASCII mode
IBM has an initiative underway to port various open source projects to
work on z/OS, running under ASCII.  But its native character set is IBM
1047.  This requires a bunch of fancy footwork that they have
undertaken.  One of the necessary things is to be able to set the
character set of a file descriptor.  The __setfdccsid() function does
that, but it and its header file, zos.h, are only defined in ASCII
builds.  This commit allows perl to work in both modes.