op_dump(): handle SVOPs and METHOPs separately
Currently SVOPs and METHOPs are dumped using the same case branch within
op_dump(). This works at the moment because the op_sv field and the
op_meth_sv field happen to occupy the same offset within their
respective structs.

This commit separates out the handling of those two OP classes, which will
also allow another commit shortly to handle METHOD ops more specifically.

OP_COREARGS is also added as another dumpable SVOP type (i.e. an op
which may have an SV hanging off op_sv).

This commit also makes it so that it now always displays the value of the
op_sv field of SVOPs, which on threaded builds starts off holding an SV,
but later gets set to 0 to indicate that the SV has been moved to the
pad.