MY_CXT: don't SEGV on a 1-byte struct.
The MY_CXT mechanism allows XS code to declare a 'static' struct
which is actually per-interpreter. Behind the scenes, the memory for
this struct is allocated as the PVX buf of an SV. If the size of the
struct is 1 (e.g. '{ char foo }' ) then newSV(size-1) gets called
as newSV(0), which skips allocating a PVX buffer. SEGVs ensue.