ParseXS: refactor: make build_subclass() simpler
The $build_subclass->() method called in BEGIN by subclasses to create
a subclass, has as the first two parameters: a key which must be
'parent', and the suffix of the parent class name. If the latter is '',
'ExtUtils::ParseXS::Node' is assumed.

This commit makes these two args optional; if not present, it acts like
('parent' => '') used to.

In addition, the 'parent' arg has been renamed to '-parent' to
distinguish it from a possible field name.

Based on a suggestion from Tony C.