ParseXS: disallow const on non-C++ methods
It's possible for an XSUB's name to include a class, in which case the
XS parser will treat it as a C++ method and declare a THIS variable.
It's also possible for an XSUB to have a postfix 'const' modifier to
mimic similar in C++ method declarations. For example this XS declaration:

    void
    X::Y::foo(...) const

causes this C++ variable to be declared:

    const X__Y * THIS = ...;

The 'const' is supposed to be an optional modifier to a C++ XSUB but the
parser currently allows it it on plain XSUBs, e.g.

    void
    foo(...) const

which leads to 'uninit var' warnings during parsing, and to badly-formed
C code or weird errors about missing typemap entries.

This commit makes it an error to have 'const' on a non-C++ XSUB.