ParseXS: split out C++ class and const
An XSUB can be a C++ method if its name includes a class; it can also
have an optional 'const' postfix modifier; e.g.:

    int
    X::Y::foo(...) const

Before this commit, both the class of the XSUB and the const modifier
were combined into a single string that was used in various places, suvh
as to declare the type of the THIS variable, e.g.

    const X__Y THIS = ...;

This commit splits the state into two separate fields during parsing,
then these two fields are used later during code generation to Do the
Right Thing.

So before, the XSUB declaration above would, at parse time, have created
an xsub_decl node which looked like:

    $xsub->{xsub_decl}{class} = 'const X::Y';

and which now looks like:

    $xsub->{xsub_decl}{class}    = 'X::Y';
    $xsub->{xsub_decl}{is_const} = TRUE;

The three tests added by this commit all failed before this fix.