PerlIO-via: fix PROTOTYPES typo
This line:

    PROTOTYPES: ENABLE;

is actually an error: such declarations shouldn't end with a semicolon.
However, the XS parser has historically treated anything it doesn't
recognise following the PROTOTYPES: keyword as if was 'DISABLE'.

So this commit changes it to  'DISABLE'. This should make no functional
difference, but it will avoid this module starting to warn when ParseXS
is updated shortly to be stricter.