ParseXS: handle PROTOTYPES keyword more strictly
See the previous commit for a general discussion on the problems
with ENABLE/DISABLE-valued keywords.

This commit specifically updates how the PROTOTYPES: keyword is handled.

Compared with the previous production release, PROTOTYPES is now
stricter in validation. It still accepts ENABLE or DISABLE in a
case-insensitive matter, but now treats any trailing text as an error,
apart from possibly a 'D' or ':'. So Anything apart these forms are
compile-time errors now:

    ENABLE
    enable
    Enabled
    DISABLE;
    DisablED;
    etc

For backwards compatibility it still treats anything apart from
/^ENABLE/ as false; so all except the first example above are treated
as DISABLE. But now, it warns if the keyword has been accepted but isn't
one of ENABLE/DISABLE.