perlapi/perlintern: Note binary-compatibility-only entries
Some API elements exist only so that already-existing uses of them can
continue to use them without change.  Generally newer/shinier methods
have come along to accomplish the tasks these do, and new code should use
those newer methods.

It effectively is deprecated to use these for new code.

This commit changes perlapi and perlintern to highlight such elements as
not being wise to use going forward.  Also, we should feel no pressure
to add documention on how to use these elements.  The code that already
uses them probably already knows how to use them.  And if it isn't using
them properly, it would be advantageous to convert to use the shinier
method, instead of trying to figure out how to get the old one to work.

However, this change shows that some elements are wrongly marked as
being binary-compatibility only.  The list of these needs to examined,
and other commits made to remove that marking when appropriate.  The
list is contained in

http://nntp.perl.org/group/perl.perl5.porters/269936