Clean up PL_strtab hash with NULL in perl_destruct
As it's not required to pre-allocate PL_strtab hash before
perl_construct() and PL_strtab is not initialized with NULL by
default, SIGSEGV could happen in certain circumstances.

Particularly, when application links libperl and runs perl several
times.  After the first run PL_strtab remains unitialized in
perl_construct() as it was not cleared in perl_destruct() and it
contains garbage and the next initialization in perl_construct() is
skipped because !PL_strtab check is skipped.

See also: c82f4881ef (Allow custom PL_strtab hash in perl_construct.)