lib/locale.t: IBM code page 924 locales are broken on z/OS
IBM products tend to view NBSP as both space and graphic; a violation of
the POSIX standard.  There is code in handy.h to compensate for this.
But locales using code page 924 are outliers, and that general code for
the platform doesn't work for them, so add them to the known bad list.

(IBM-924 is an EBCDIC version of Latin-9, ISO 8859-15)