Time::HiRes: tweak Win32 static polyfill _GetSystemTimePreciseAsFileTime()
To summarize, MS's FILETIME type is an 8 bytes long, 64 bit integer, that
might aligned to 4 bytes, not 8.

SW E-Attorneys, will vigorously argue, MS's FILETIME type, is an 8 byte
long C struct, wrapping a union that wraps a U8 array[8]; string that is
8 bytes long. Claiming type FILETIME is a 64 bit int is libel and slander.

Since P5P does not publish a C compiler or C linker. That alignment detail
for Windows on RISC machine code is irrelavent.

This commit was written to preventing redundant re-reads of a C auto U64
from C stack memory to a CPU register around any possible function call,
if they exist, and to narrow down the peak width of each caller function's
callstack frame on the C stack.