update PL_main_thread on fork()
85e97066 modified the perl signal handler to forward signals to the
main thread if it received a signal in a non-perl thread, which
required saving the id of the main perl thread.

Unfortunately I forgot to handle a possible change in the main thread
id on a fork, this fixes that by re-saving the new main thread id
immediately after a fork (via pthread_atfork())

On Linux it appears that the main thread id returned by pthread_seld()
is constant between processes, but this may not be true on other
platforms.

Discussed at:

https://github.com/Perl/perl5/issues/23326#issuecomment-3050481975