Rename _isQUOTEMETA to isQUOTEMETA
The former is undefined behavior in C in some situations.
The rules are detailed in "Choosing legal symbol names" in perlhacktips

This commit omits the underscore and makes the macro available only to
the core and extensions.  I think I didn't know about PERL_CORE when I
created this macro, so used the leading underscore to discourage its
use from anyone who stumbled upon its existence.  The reason not to make
it public was my uncertainty about if it was the correct thing to do;
my not expecting that it would be useful outside of core; and the
extra work needed to make it a polished interface.  I think now that if
someone requested it be made public, that could safely be done.