Add ability to define a simple implementation in embed.fnc
This can replace most of mathoms.c and lots of essentially boiler plate
scattered throughout the code.

Quite a few of our functions and macros are specializations of more
general ones, taking fewer parameters than the generalized ones.  Often
we will have a
    foo_flags(a, flags)
and a
    foo(a)
which just calls foo_flags with a flag of 0.

This commit allows for a single extra line to be added to embed.fnc to
indicate the implementation of plain foo, replacing the extra boiler
plate lines that were needed to do this prior to this commit.

Another advantage is this lowers the likelihood of the implementation of
long-name functions getting out-of-sync with supposedly-equivalent short
name macros that are generally used to bypass them.  This commit causees
the long name form to be automatically generated, so there is no
possibility of getting out-of-sync.

It uses the current infrastructure in the embed regeneration to
automatically create any needed long names, inserting thread context if
necessary without the programmer needing to worry about this.  The
results are placed in embed.h and proto.h

I'm not thrilled about the syntax of this.  You add an extra
continuation line to the declaration.  This line begins with an '=',
followed by the implementation details.  For example

 Adp    |U8 *   |uv_to_utf8     |NN U8 *d   \
				|UV uv	    \
              = uv_to_utf8_flags(d,uv,0)

But embed.fnc is the file that already generates embed.h and proto.h,
and no extra information beyond what embed.fnc already contains needs to
be added.

The bottom line is that many few-line functions in our code can be
replaced by adding a single line to embed.fnc, which will make those
into macros instead of functions.  This requires fewer resources; the
generated macros require no resources unless actually called; and it's
simply more convenient to do this here in one place, than do the extra
boiler plate that has to be in multiple places otherwise.

embed.h will now contain long name definitions that formerly it didn't,
so autodoc.pl has to change to look for them

This commit changes nothing; just adds the capability.  The next few
commits will start to use it.