Use new embed.fnc implementation capability for sv_utf8_downgrade
This shows how taking a #define for sv_utf8_downgrade() from sv.h and
moving it to embed.fnc creates the equivalent entry in embed.h, plus
automatically generating macros that implement the long form
Perl_sv_utf8_downgrade().  This means the manual implementation of
Perl_sv_utf8_downgrade() in mathoms.c can be and should be deleted,
which this commit also does.