Change #define Perl_foo  Perl_bar" to plain "foo bar"
This reverts to the naming that prevailed until 5.41:

    #define foo  bar

The  Perl_ prefixes were introduced in expectation that a scheme would
be used to allow a bunch of functions to be converted to macros.  This
actually didn't work, so was reverted by GH #23209.

However those #define changes didn't need to be reverted, so weren't

A revised scheme is now being introduced in the next commit, which needs
the original #define forms.  So this commit reverts back to those.