Allow automatic long name macro generation
Prior to this commit, a macro that doesn't deal with thread context can
have its long 'Perl_foo' name form be automatically generated.  This was
added in 5.41 by GH #22691.  I had thought it worked in all cases, but
the threaded case had to be reverted by #23209.

The original scheme was too simplistic.  This new commit reintroduces
the thread context automatic handling, but isn't so naive this time.

This commit will enable the emptying of much of mathoms.c, and the
removal of quite a few similar functions scattered throughout the code.

The next commit will convert the first such function, as a proof of
concept.