mathoms.c: Convert Perl_foo definitions to call foo
Unless there is some reason not to, a mathoms Perl_ function should just
call the macro it is the long name for.

That way, when the macro changes, the function automatically
correspondingly does too.

I discovered the hard way that one reason not to change is if the macro
expands differently depending on where in the code it is called from.
sv_setsv() always calls sv_setsv_flags(), but one of the flags it passes
is 0 when not called from the perl core.  When mathoms is compiled, we
don't know the caller, so the code assumes it isn't called from core
(which it isn't).

This commit changes all the functions in mathoms whose definitions
expand exactly to the macro they are supposed to wrap, into calling that
macro instead of duplicating the expansion in their mathoms.c bodies.

I checked this by having cpp create mathoms.i and then I saved it under
a different name.  Then I changed all the functions to call their
implementing macro.  Then I ran cpp again to create a revised mathoms.i.
Then I compared the two .i files.  I backed out the changes to those
functions whose expansions differed between the two files.  The result
is this commit