foreign thread signal forwarding: swap tests to support OpenBSD
On OpenBSD, With the tests in the original order, with the parent
process setting a signal handler and accepting a foreign signal, the
child process test would block busy waiting when the child process
called exit.

From looking at kdump the child process is busy waiting on sched_yield
outside the perl process.

Since this appears to be OpenBSD specific (other BSD, linux, MacOS
work fine), and the sequence causing the problem seems unlikely
outside a test scenario I'm not going to investigate further.