embed.fnc: Need p flag for non-static public elements
If a function or macro 'foo' is publicly accessible, there should be a
'Perl_foo' form available for callers to use so as to avoid any name
collisions with their own names.

Now that we can automatically generate Perl_foo equivalents for macros,
make sure that the flags indicate that.

I first thought to add the p flag automatically, but that would require
a bunch of restructuring.  By requiring the flag, in embed.fnc, the true
state of things doesn't get hidden.