Rm Perl_sv_pvutf8() from mathoms
This function can be converted to call plain sv_pvutf8  without changing
its behavior.  I verified this by eyeballing its implementation in
mathoms.c and comparing that with what happens with just the plain
macro.

That means embed.fnc can be changed to generate the Perl_ form
automatically, and the mathoms function can be removed.