Add Perl_expected_size stub macro
Passed a number of bytes, this macro will return:
* PERL_STRLEN_NEW_MIN if that is larger than the argument value
* the argument value rounded up to PTRSIZE

The intention is to later convert `Perl_expected_size` into a
function that has some knowledge of the underlying malloc library's
behaviour and can return values more reflective of the actual
useable allocation size for the specified number of bytes.

This can aid in reducing unnecessary calls to realloc().