Revert "Attempt to remove deprecated sv_locking/unlocking"
This reverts commit 3afafdb6cda45b8879aa284d5fb90c3446584bfc.

It turns out that Perl_sv_nounlocking() is the mechanism that is used to
make sure mathoms.c gets loaded, as the comments in the reverted commit
say, but which I had not noticed.

Without this function being in mathoms.c, some build configurations will
not load mathoms, leaving other functions in mathoms.c undefined, as
happened with the two issues this fixes.

This commit could be reinstated if another method is created to force
mathoms to be loaded;

This fixes #23473
This fixes #23474