Reapply "Attempt to remove deprecated sv_locking/unlocking"
This reverts commit f3a5b2b22f875a08c4d1253a866327c4551f0d73.

That commit reverted 3afafdb6cda45b8879aa284d5fb90c3446584bfc.  The
reason was one of the removed deprecated functions being present in
mathoms had an extra purpose, making sure that the functions in
mathoms.c get loaded.

The commit just prior to this one created a different mechanism for
forcing that load, so the the reason for the original revert is gone, so
it can be reapplied.