Silence autodoc warning
This was recently introduced in 194d4eac2f635f5c4cfc1e3709146be8063d6773
The warning was that IS_SAFE_SYSCALL has no long name.  We currently
don't typically have long names for macros.  The solution is to mark it
as a macro; autodoc thought it must be a function.

Spotted by Dave Mitchell.