toke.c: Rm use of E<> on apidoc lines
These lines are used for various downstream tools, only one of which so
far has need to know about E<> pod markup, and confuses the others.  The
outlier tool is autodoc.pl, so change that to add the E<> when needed.

The code to do this (more generalized than the immediate need here) was
copied from Dan Book, modified by a comment from Lukas Mai.